GFETCH-116 moved LocationCallback, LocationService to location package
