+++ app/src/main/java/co/gobd/tracker/ui/MainActivity.java
@@ -15,7 +15,7 @@ import android.widget.Button;
 import android.widget.Toast;
 
 import co.gobd.tracker.R;
-import co.gobd.tracker.service.LocationService;
+import co.gobd.tracker.service.location.LocationService;
 
 public class MainActivity extends AppCompatActivity implements View.OnClickListener {
